fix: runtime init panic
