+++ examples/plugin/main.go
@@ -1,16 +1,18 @@
+//go:build native
+// +build native
+
 package main
 
 import (
 	"fmt"
 
-	"kcl-lang.io/kcl-go/pkg/kcl"
-	"kcl-lang.io/kcl-go/pkg/native"                // Import the native API
+	"kcl-lang.io/kcl-go/pkg/kcl"                   // Import the native API
 	_ "kcl-lang.io/kcl-go/pkg/plugin/hello_plugin" // Import the hello plugin
 )
 
 func main() {
 	// Note we use `native.MustRun` here instead of `kcl.MustRun`, because it needs the cgo feature.
-	yaml := native.MustRun("main.k", kcl.WithCode(code)).GetRawYamlResult()
+	yaml := kcl.MustRun("main.k", kcl.WithCode(code)).GetRawYamlResult()
 	fmt.Println(yaml)
 }
