+++ kclvm_test.go
@@ -312,17 +312,11 @@ func TestGetSchemaType(t *testing.T) {
 			SchemaName: "Person",
 			Properties: map[string]*gpyrpc.KclType{
 				"name": {
-					Type:       "str",
-					Line:       1,
-					Properties: map[string]*gpyrpc.KclType{},
-					Required:   []string{},
-					UnionTypes: []*gpyrpc.KclType{},
-					Decorators: []*gpyrpc.Decorator{},
-					Examples:   map[string]*gpyrpc.Example{},
+					Type: "str",
+					Line: 1,
 				},
 			},
-			Required:   []string{"name"},
-			UnionTypes: []*gpyrpc.KclType{},
+			Required: []string{"name"},
 			Decorators: []*gpyrpc.Decorator{
 				{
 					Name:      "info",
@@ -332,7 +326,6 @@ func TestGetSchemaType(t *testing.T) {
 					},
 				},
 			},
-			Examples: map[string]*gpyrpc.Example{},
 		},
 	}, result)
 	p, err := filepath.Abs("./testdata/main.k")
@@ -351,30 +344,17 @@ func TestGetSchemaType(t *testing.T) {
 			SchemaName: "Person",
 			Properties: map[string]*gpyrpc.KclType{
 				"name": {
-					Type:       "str",
-					Line:       1,
-					Default:    "\"kcl\"",
-					Properties: map[string]*gpyrpc.KclType{},
-					Required:   []string{},
-					UnionTypes: []*gpyrpc.KclType{},
-					Decorators: []*gpyrpc.Decorator{},
-					Examples:   map[string]*gpyrpc.Example{},
+					Type:    "str",
+					Line:    1,
+					Default: "\"kcl\"",
 				},
 				"age": {
-					Type:       "int",
-					Line:       2,
-					Default:    "1",
-					Properties: map[string]*gpyrpc.KclType{},
-					Required:   []string{},
-					UnionTypes: []*gpyrpc.KclType{},
-					Decorators: []*gpyrpc.Decorator{},
-					Examples:   map[string]*gpyrpc.Example{},
+					Type:    "int",
+					Line:    2,
+					Default: "1",
 				},
 			},
-			Required:   []string{"name", "age"},
-			UnionTypes: []*gpyrpc.KclType{},
-			Decorators: []*gpyrpc.Decorator{},
-			Examples:   map[string]*gpyrpc.Example{},
+			Required: []string{"name", "age"},
 		},
 	}, result)
 }
@@ -391,19 +371,11 @@ func TestGetSchemaTypeMapping(t *testing.T) {
 		SchemaName: "Person",
 		Properties: map[string]*gpyrpc.KclType{
 			"name": {
-				Type:       "str",
-				Line:       1,
-				Properties: map[string]*gpyrpc.KclType{},
-				Required:   []string{},
-				UnionTypes: []*gpyrpc.KclType{},
-				Decorators: []*gpyrpc.Decorator{},
-				Examples:   map[string]*gpyrpc.Example{},
+				Type: "str",
+				Line: 1,
 			},
 		},
-		Required:   []string{"name"},
-		UnionTypes: []*gpyrpc.KclType{},
-		Decorators: []*gpyrpc.Decorator{},
-		Examples:   map[string]*gpyrpc.Example{},
+		Required: []string{"name"},
 	}
 	assert2.Equal(t, map[string]*gpyrpc.KclType{
 		"Person": &personSchema,
@@ -415,28 +387,15 @@ func TestGetSchemaTypeMapping(t *testing.T) {
 			BaseSchema: &personSchema,
 			Properties: map[string]*gpyrpc.KclType{
 				"name": {
-					Type:       "str",
-					Line:       1,
-					Properties: map[string]*gpyrpc.KclType{},
-					Required:   []string{},
-					UnionTypes: []*gpyrpc.KclType{},
-					Decorators: []*gpyrpc.Decorator{},
-					Examples:   map[string]*gpyrpc.Example{},
+					Type: "str",
+					Line: 1,
 				},
 				"count": {
-					Type:       "int",
-					Line:       2,
-					Properties: map[string]*gpyrpc.KclType{},
-					Required:   []string{},
-					UnionTypes: []*gpyrpc.KclType{},
-					Decorators: []*gpyrpc.Decorator{},
-					Examples:   map[string]*gpyrpc.Example{},
+					Type: "int",
+					Line: 2,
 				},
 			},
-			Required:   []string{"count", "name"},
-			UnionTypes: []*gpyrpc.KclType{},
-			Decorators: []*gpyrpc.Decorator{},
-			Examples:   map[string]*gpyrpc.Example{},
+			Required: []string{"count", "name"},
 		},
 	}, result)
 
@@ -456,30 +415,17 @@ func TestGetSchemaTypeMapping(t *testing.T) {
 			SchemaName: "Person",
 			Properties: map[string]*gpyrpc.KclType{
 				"name": {
-					Type:       "str",
-					Line:       1,
-					Default:    "\"kcl\"",
-					Properties: map[string]*gpyrpc.KclType{},
-					Required:   []string{},
-					UnionTypes: []*gpyrpc.KclType{},
-					Decorators: []*gpyrpc.Decorator{},
-					Examples:   map[string]*gpyrpc.Example{},
+					Type:    "str",
+					Line:    1,
+					Default: "\"kcl\"",
 				},
 				"age": {
-					Type:       "int",
-					Line:       2,
-					Default:    "1",
-					Properties: map[string]*gpyrpc.KclType{},
-					Required:   []string{},
-					UnionTypes: []*gpyrpc.KclType{},
-					Decorators: []*gpyrpc.Decorator{},
-					Examples:   map[string]*gpyrpc.Example{},
+					Type:    "int",
+					Line:    2,
+					Default: "1",
 				},
 			},
-			Required:   []string{"name", "age"},
-			UnionTypes: []*gpyrpc.KclType{},
-			Decorators: []*gpyrpc.Decorator{},
-			Examples:   map[string]*gpyrpc.Example{},
+			Required: []string{"name", "age"},
 		},
 	}, result)
 }
